[frontend/backend] explanations on analyses tab
