+++ opencti-platform/opencti-front/lang/front/en.json
@@ -2584,7 +2584,7 @@
   "Pipe": "Pipe",
   "PIR": "PIR",
   "PIR Details": "PIR Details",
-  "PIR Rescan period (days)": "PIR Rescan period (days)",
+  "Rescan period (days)": "Rescan period (days)",
   "Place of Birth": "Place of Birth",
   "Place of birth": "Place of birth",
   "Plain text": "Plain text",
@@ -2611,6 +2611,7 @@
   "Platform title": "Platform title",
   "Platform under construction, subscribe to update!": "Platform under construction, subscribe to update!",
   "Platforms": "Platforms",
+  "Entities in PIR": "Entities in PIR",
   "Playbook": "Playbook",
   "Playbook is running": "Playbook is running",
   "Playbook is stopped": "Playbook is stopped",
