+++ opencti-platform/opencti-front/src/private/components/pir/Pir.tsx
@@ -11,7 +11,7 @@ import PirOverview from './pir_overview/PirOverview';
 import ErrorNotFound from '../../../components/ErrorNotFound';
 import useQueryLoading from '../../../utils/hooks/useQueryLoading';
 import Loader from '../../../components/Loader';
-import PirAnalyses from './PirAnalyses';
+import PirAnalyses from './pir_analyses/PirAnalyses';
 
 const pirQuery = graphql`
   query PirQuery($id: ID!) {
