+++ opencti-platform/opencti-graphql/config/schema/opencti.graphql
@@ -3380,6 +3380,7 @@ interface Container {
     search: String
     types: [String]
     all: Boolean
+    pirId: String
   ): StixObjectOrStixRelationshipRefConnection
   relatedContainers(
     first: Int
@@ -3540,6 +3541,7 @@ type Note implements BasicObject & StixObject & StixCoreObject & StixDomainObjec
     search: String
     types: [String]
     all: Boolean
+    pirId: String
   ): StixObjectOrStixRelationshipRefConnection
   relatedContainers(
     first: Int
@@ -3773,6 +3775,7 @@ type ObservedData implements BasicObject & StixObject & StixCoreObject & StixDom
     search: String
     types: [String]
     all: Boolean
+    pirId: String
   ): StixObjectOrStixRelationshipRefConnection
   relatedContainers(
     first: Int
@@ -3982,6 +3985,7 @@ type Opinion implements BasicObject & StixObject & StixCoreObject & StixDomainOb
     search: String
     types: [String]
     all: Boolean
+    pirId: String
   ): StixObjectOrStixRelationshipRefConnection
   relatedContainers(
     first: Int
@@ -4209,6 +4213,7 @@ type Report implements BasicObject & StixObject & StixCoreObject & StixDomainObj
     search: String
     types: [String]
     all: Boolean
+    pirId: String
   ): StixObjectOrStixRelationshipRefConnection
   relatedContainers(
     first: Int
