+++ opencti-platform/opencti-graphql/src/modules/case/case-rfi/case-rfi.graphql
@@ -114,6 +114,7 @@ type CaseRfi implements BasicObject & StixObject & StixCoreObject & StixDomainOb
     search: String
     types: [String]
     all: Boolean
+    pirId: String
   ): StixObjectOrStixRelationshipRefConnection
   relatedContainers(
     first: Int
