+++ opencti-platform/opencti-graphql/src/modules/case/case.graphql
@@ -109,6 +109,7 @@ interface Case implements BasicObject & StixObject & StixCoreObject & StixDomain
     search: String
     types: [String]
     all: Boolean
+    pirId: String
   ): StixObjectOrStixRelationshipRefConnection
   relatedContainers(
     first: Int
