+++ opencti-platform/opencti-graphql/src/modules/case/feedback/feedback.graphql
@@ -109,6 +109,7 @@ type Feedback implements BasicObject & StixObject & StixCoreObject & StixDomainO
     search: String
     types: [String]
     all: Boolean
+    pirId: String
   ): StixObjectOrStixRelationshipRefConnection
   relatedContainers(
     first: Int
