+++ opencti-platform/opencti-graphql/src/modules/pir/pir-resolvers.ts
@@ -9,7 +9,7 @@ const pirResolvers: Resolvers = {
   Pir: {
     creators: (pir, _, context) => context.batch.creatorsBatchLoader.load(pir.creator_id),
     objectMarking: (pir, _, context) => context.batch.markingsBatchLoader.load(pir),
-    pirContainers: (pir, args, context) => findPirContainers(context, context.user, pir.id, args),
+    pirContainers: (pir, args, context) => findPirContainers(context, context.user, pir, args),
   },
   Mutation: {
     pirAdd: (_, { input }, context) => pirAdd(context, context.user, input),
