+++ opencti-platform/opencti-graphql/src/modules/task/task.graphql
@@ -108,6 +108,7 @@ type Task implements Container & StixDomainObject & StixCoreObject & StixObject 
     search: String
     types: [String]
     all: Boolean
+    pirId: String
   ): StixObjectOrStixRelationshipRefConnection
   relatedContainers(
     first: Int
